ref: Move Cacher directly into actors
