+++ src/actors/mod.rs
@@ -1,5 +1,5 @@
+pub mod cacher;
 pub mod cficaches;
-pub mod common;
 pub mod objects;
 pub mod symbolication;
 pub mod symcaches;
